Added all others repository

+++ src/main/java/com/example/pizza/repositories/InformationRepository.java
@@ -3,5 +3,5 @@ package com.example.pizza.repositories;
 import com.example.pizza.model.Information;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InformationRepository extends JpaRepository<Information, Integer> {
+public interface InformationRepository extends JpaRepository<Information, Long> {
 }
